[IMP] l10n_uk: HMRC connection improvement
Contains four improvements for HMRC flow:
1- Successfully connecting to HMRC now shows a notification to let the
user know they can send their reports.
2- The "send" button in the send HMRC wizard triggers the sending but
nothing happens UI wise. Now, the wizard is closed and a confirmation
notification is shown.
3- For a user wanting to submit the VAT report for two companies, after
switching companies, clicking on the "send to HMRC" button raises an
error as the token does not correspond to the correct company. Before,
the user had to manually remove their authentication credentials to
allow for a new connection to get a new token. Now, if we get an error
from HMRC that the agent isn't authorised to submit the report, we clear
the credentials ourselves and redirect the user to the login page, so
they could get the correct token.
4- Trying to send the report without selecting an obligation returns a
traceback error. Now, the send button will not appear unless an
obligation is selected.

task-3696341

closes odoo/enterprise#60375

X-original-commit: d5d4bbcc2200d3826e9677190580c34631683485
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>