[FIX] account_winbooks_import: reduce memory footprint
Importing large winbooks zip files can sometimes lead to a MemoryError.
For instance, importing 360 000 account.moves in one go requires
3.1Gb of allocated memory, too much for a 2gb worker.

This commit removes the intermediate variable recs to avoid storing
the result of DBF.read in memory. This intermediate variable prevented
dbf_records to be garbage-collected before the end of the
_import_moves function. This makes the peak memory drop to 2.4gb.
Also changes the recfactory of DBF to frozendict.
As frozendicts are hashable they can be added to a set to remove
duplicates. By removing the call to Orderdict.items(),
the memory peak drops further to 1.6Gb, allowing to import the winbooks
export on a 2gb worker.

Add a more user-friendly error when there is no sub_zips in the main
zip container. This can happen if the customer extracts the winbooks
archive and tries to import the sub_zip directly.

opw-3634244

closes odoo/enterprise#59371

X-original-commit: 1386404ac8c48b615d506343b84d25ebe7222e3b
Signed-off-by: William André (wan) <wan@odoo.com>