[FIX] l10n_account_customer_statement: fix auto sending start date.
At the moment, when auto sending the customer statement, it will
always send the values for the current month.

This means that if you send it bi-weekly, the first sending will contain
the values for the first two weeks of the month, while the second will
contain the values for the whole month.

But this does not work well when you can have custom frequency, as there
will be cases, where you will want the values of the last x weeks/months,
or when a two-week period is between two months.

To have a more flexible system, we will now compute the start date by taking
into account the set frequency. The end date of the statement will always be
the day before the sending date, while the start date will be the end date minus
the frequency (so minus two weeks if set to bi-weekly).

Task id # 3713352

closes odoo/enterprise#57723

X-original-commit: 14ff43630be03b1b54437d3ab00c74ff3a8cf581
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>