[FIX] documents: fix traceback caused by sample data
Bug
===
When we show the sample data of documents, it will choose a random
value for the selection field. But document use a special option
to make the browser generate the thumbnail of PDFs and then save it
on the record. So, without this rule, it will try to update the
record if "client_generated" has been chosen.

Task-3697924

closes odoo/enterprise#56527

X-original-commit: bf15c2a0e4c28d587bfaae9ef1a4af041f835446
Related: odoo/odoo#153876
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>