[IMP] website_sale_{*}, *: move delivery selection to /checkout page
A single step for selecting a delivery method and a payment provider
poses two problems:
1) The interface looks cluttered when multiple pms and delivery methods
are configured.
2) Complexity in a payment and a delivery widget dependencies. As they are
rendered on the same page it requires to dynamically update parts of its
content in js causing bugs and inconsistencies.

For those reasons we decided to move the delivery selection step to
the /checkout step where an address is selected.

task-3133415

closes odoo/enterprise#56049

Related: odoo/upgrade#6035
Related: odoo/odoo#153063
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>