[FIX] web_studio: only add studioXpath to pages in notebook
This commit fixes the notebook compiler of the form editor. Since some views
define invisible fields directly inside a notebook, the compiler was wrong
when trying to match the xpath of a notebook tab with the corresponding node
from the compiled Notebook component.

This created issues, such as incorrect or null xpaths, preventing to edit some
tabs and pages, or crashing when adding an element.

opw-3672498

closes odoo/enterprise#56495

X-original-commit: 48ad544b48e4363a02c74376c3fed34fe759898e
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>