[IMP] account_accountant: defer on accounting date
When we use the mode `on_validation` for deferred entries, the first
entry that puts the total amount on the deferred account is happening
at the end of the month of the move.

When we have a vendor bill on Feb 1, 2024, and the amount is deferred
from Feb 1, 2024 to Mar 31, 2024, we will have the following moves:

[Feb 29, 2024] $1000 from Expenses to Deferred Expenses
[Feb 29, 2024]  $500 from Deferred Expenses to Expenses
[Mar 31, 2024]  $500 from Deferred Expenses to Expenses

That means that during the whole month of February we have the total
amount on our Expenses account, except on the last day.

In order to improve this, we now defer the original move on the same
accounting date of the move, such that during that month at most the
deferred part is on the Revenue/Expense account (on the last day).

task-3725621

closes odoo/enterprise#56554

X-original-commit: fcb646ce4b29dc2215e17533f0117d315f267ec4
Related: odoo/documentation#7729
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>