[FIX] hr_payroll: avoid test dependency on sign
Before this commit, the `test_dashboard_ui` test
in payroll was relying on sign.
However, `hr_payroll` does not depend on sign,
which would lead to the test breaking when that
module is not installed.

This commit adds a check to see if sign is installed
before creating the additional data related to it.

closes odoo/enterprise#56300

X-original-commit: 43a04f2b619bba1283e9a5129866f0a7ecfc4824
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>