[FIX] hr_payroll: avoid test dependency on sign
Before this commit, the `test_dashboard_ui` test
in payroll was relying on sign.
However, `hr_payroll` does not depend on sign,
which would lead to the test breaking when that
module is not installed.

This commit adds a check to see if sign is installed
before creating the additional data related to it.

closes odoo/enterprise#56111

X-original-commit: 6b914e54193828be14aee989c49fca58f3760aeb
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>