[FIX] l10n_nl_reports_sbr: tests were still standard
The test classes should not be standard: it makes the test run in l10n builds, while we only expect them to be tested on external builds.
So, we can have issues with the server from the government not responding which causes errors on l10n builds

It also causes issues in 16.4+ where the request is not accepted.

Linked to runbot error 23358, 57019

closes odoo/enterprise#56471

X-original-commit: 0efb87c4b474d50d81b8c189110e9dcefc2161b8
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>