[IMP] knowledge,website_knowledge: improve front-end sub-sites
This commit improves the front-end part of Knowledge which was until now a bit
confusing.
The old version of the front-end displayed all of the published articles which
resulted in the view being really cramped, and a public user could access
unrelated articles to the one he was given the link for.

When a user publishes an article we now automatically publish its children in
order to create subtrees containing a part of the published articles.
This way sub-sites are created for the public users, when they access an
article via the specific `/knowledge/article/<article_id>`.
Only published articles that are below a root, which is the highest published
ancestor are thus visible on the front-end.

The query counter of the 'test_article_home_page' was adapted as we changed
the way the 'first accessible article' is retrieved to also take into account
visible articles.

task-3607868

closes odoo/enterprise#56122

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>