[FIX] web_studio: subview's button_box doesn't pollute its main view's template
Have a form view, with a subview form.
That subview has a button_box.

Before this commit, the button box was compiled in the context of the main view, eventually
leading to crashes.

After this commit, it is compiled only when navigation to that subview for edition.

This issue and the fix are similar to odoo/odoo#149896

opw-3681966

closes odoo/enterprise#56245

X-original-commit: b1822c22e77af6b9d765a4f278ad3416566b7a5e
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>