[IMP] *: Upgrade Bootstrap to version 5.3
*: account_followup, account_reports, documents, frontdesk, room, sign,
   snailmail_account_followup, web_studio

This commit upgrades Bootstrap library from versison 5.1.3 to 5.3.3.

In a nutshell, the big changes are :
- A lot more CSS custom properties all over the place
- ScrollSpy has been completely rewritten (moving from onscroll listener
  to IntersectionObserver)
- Support for color modes (aka. light/dark modes ; but currently
  disabled as we don't take advantage of it yet).

Note: even though the new color modes have a lot of potential, as we
don't take advantage of it yet, the built-in dark mode is currently
disabled. It should be enabled once actually implemented.

This commit also removes the `.bg-text-*` helpers backported from
Bootstrap 5.2 (and now built-in) and an override of the
`Tooltip.prototype.setElementContent()` function as the Tooltip
component now delegates the its content assignation to the
`TemplateFactory` class.

References:
- https://getbootstrap.com/docs/5.3/migration

task-3600468

Part-of: odoo/enterprise#59122