[PERF] mrp_workorder: avoid recomputing barcode target record
When a user opens the Shop Floor app, each MrpDisplayRecord will compute the barcode target record based on the admin ID.
This is not necessary as the barcode target record will be the same as long as the admin ID does not change.
This can lead to slow computations and make the browser crash when there are many records.

Before this commit, all MrpDisplayRecord were recomputing the barcode target record.

After this commit, we cache the admin ID and the barcode target record ID to avoid recomputing it if unecessary.

opw-3721896

closes odoo/enterprise#56167

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>