[FIX] account_reports: tax tags engine: don't return duplicate empty result in case of shared tag
It is possible that two expressions (possibly of different report, linked together by a cross_report expression) needing to be evaluated at the same time share the same tax tags formula. When this case happens, the result of the engine function should be in the form {(formula, recordset_of_expr_1_and_2): result} (assuming there is no other expression to evaluate, for the sake of simplicity).

Before this fix, in this situation, the engine instead returned:
{
    (formula, expr_1): result,
    (formula, expr_2): result,
    (formula, recordset_of_expr_1_and_2): empty_value,
}

Because of that, the report ended up evalutating these expressions with an empty result. This was spotted when forward-porting https://github.com/odoo/enterprise/commit/c8beb3559b9199fc2f72773dd5dedf871cf2406b , since the new error it introduces raised in such case, since the value initially computed for expr_1/expr_2 got overwritten.

This was wrong, and was due to the fact the result dict was first initialized using the full key of the formulas_dict (hence grouping expr_1 and expr_2 into a common recordset), and then wrongly iterated on all the expressions contained in the keys. We fix that by using the full key everywhere.

closes odoo/enterprise#56217

X-original-commit: e41d2ce549b4541d5b8745c3cfb53b59de15ed94
Signed-off-by: John Laterre (jol) <jol@odoo.com>