[FIX] pos_blackbox_be: fix control buttons
In the previous commit, an issue arose where the point of sale became unusable after installing the pos_blackbox_module.
This issue raised from an omission in updating the code for control buttons, specifically within the Belgian module.
With this commit, we ensure that the necessary adjustments are made to rectify this issue and restore the functionality of the point of sale for Belgian users that required the blackbox.

closes odoo/enterprise#56164

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>