[FIX] pos_preparation_display: sync on reconnect
Current behavior:
When the websocket connection is lost, the webworker will attempt to
reconnect. However, the messages that were sent in the meantime are
lost.

The fix:
The websocket webworker attempts reconnection indefinetely, so we do
know that at some point it will indeed reconnect.
To solve this, we listen to the reconnect event and retrieve the
orders that were sent in the meantime.

opw-3701020

closes odoo/enterprise#56163

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>