[FIX] pos_preparation_display: sync on reconnect
Current behavior:
When the websocket connection is lost, the webworker will attempt to
reconnect. However, the messages that were sent in the meantime are
lost.

The fix:
The websocket webworker attempts reconnection indefinetely, so we do
know that at some point it will indeed reconnect.
To solve this, we listen to the reconnect event and retrieve the
orders that were sent in the meantime.

opw-3701020

closes odoo/enterprise#57055

X-original-commit: 647eb5286063c471c45c23d8c00316d0499b1d42
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>