[FIX] documents: fix trash button visibility
Purpose
=======
Fix the trash button which was disappearing from
the inspector when multiple documents are selected.

Specifications
==============
When only one document is selected, the trash button is
displayed next to the "Name" field in the inspector.
When multiple documents are selected, this "Name" field is
hidden preventing the trash button to be displayed.
Moving the trash button next to the "Contact" field when multiple
documents are selected for it to keep being accessible.

Task-3700965

closes odoo/enterprise#56229

X-original-commit: b341369a714fe945695e017f41d8723f1e89236a
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>