[IMP] hr_payroll_holidays: add a warning in the payroll dashboard for leaves with no allocation
When a time off for a few days is taken using a certain allocation but then the number of days of the allocation is changed, there should be a warning on the dashboard of payroll to show the leaves affected.

The warning redirects the user to a list of leaves concerned and grouped by employees.

closes odoo/enterprise#58570

Task: 3729563
X-original-commit: 1bb979291655949632c535a93af3f4c157465c75
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Arthur Pierrot (arpi) <arpi@odoo.com>