[IMP] hr_payroll_holidays: add a warning in the payroll dashboard for leaves with no allocation
When a time off for a few days is taken using a certain allocation but then the number of days of the allocation is changed, there should be a warning on the dashboard of payroll to show the leaves affected.

The warning redirects the user to a list of leaves concerned and grouped by employees.

closes odoo/enterprise#56188

Task: 3729563
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>