[FIX] hr_payroll: access error when contract company is not selected
Currently, when an employee contract belongs to a company different
than the employee's company, a warning is shown in the payroll
dashboard [1].
However, if the current user has no access to the contract's company, an
access error is raised, which makes not possible to neither see the
warning nor load the dashboard.

This commit fixes the above issue by ensuring the employee contract is
read as sudo, to avoid requiring the current user to have both companies
selected to see the warning.

References:

- [1] https://github.com/odoo/enterprise/blob/e41d2ce5/hr_payroll/models/hr_payslip.py#L1072

closes odoo/enterprise#56865

X-original-commit: 7cbdac043e6b443c85dec5e9c366aabe12ef6130
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>