[FIX] data_merge: do not inspect non-loaded models
This patch is an improvement and simplification over the fix done in
5a64974501b5be231e2eed941f42ae6dfe110f18

When generating the company search query we should skip unloaded modules
to avoid a `KeyError`. This issue became more visible now that the check
in `ir.rule` is done for each domain. Refer to the above commit for
steps on how to reproduce the issue.

This issue has a high impact in upgrades since it potentially blocks all
upgrades for which a `data_merge.record` entry exists for a custom model.

closes odoo/enterprise#56327

X-original-commit: d24ea6b1434de4700a24e548bf63f14b2ab9840b
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>