[FIX] sale_subscription: draft invoice with error payment token
Before this commit, when a customer paid a subscription from 'Generate payment link' and got an error in the transaction or if it was canceled, Odoo was creating a draft invoice (it remained open) and saving the faulty token.

After this commit. the draft invoice is canceled right after the error is detected and the token is saved only if the renewal state of the subscription is authorized.

closes odoo/enterprise#56342

Task-id: 3734247
X-original-commit: 32f6e04ec39afb5b63b4447278c058ddb6b2f6e7
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>