[FIX] documents_spreadsheet: remove unreliable test step
In #48767, we introduced the collapsing of a certain side panel feature
relying on Bootstrap.
Bootstrap apparenty has an internal "popote" to make the target node
fold and unfold and one of our tests actually tests against this internal
popote.

Unfortunately, the test is not reliable and from a computer to the other
or even depending on the browser you use, it does not behave consistently.

This commit removes the unreliable step as it does not test our own
codebase.

closes odoo/enterprise#56379

Task: 3736963
X-original-commit: ccb5804eda8c53f350004d94e6e6bc9e445af9f2
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>