[FIX] documents_spreadsheet: remove unreliable test step
In #48767, we introduced the collapsing of a certain side panel feature
relying on Bootstrap.
Bootstrap apparenty has an internal "popote" to make the target node
fold and unfold and one of our tests actually tests against this internal
popote.

Unfortunately, the test is not reliable and from a computer to the other
or even depending on the browser you use, it does not behave consistently.

This commit removes the unreliable step as it does not test our own
codebase.

closes odoo/enterprise#56291

Task: 3736963
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>