[FIX] l10n_co_reports: do not report void lines
Create a Vendor Bill with a Colombia Partner and withholding tax
Accounting > Reporting > Certificado de Retención en IVA
Print PDF

Issue: Currently the lines in the report are being generated
despite the balance in “tax_base_amount” (“Monto Total Operacion) being 0.00.

Expected behavior: If the “tax_base_amount is 0.00 do not consider this balance
neither in the report lines nor in the SUM of the Report total.

opw-3704972

closes odoo/enterprise#57089

X-original-commit: 543801f6deeb944e9fe5752f4ce6e5f34478e51e
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>