[FIX] knowledge: close dropdown on click in embedded views
Step to reproduce:

- Insert a view using /item list or /item kanban in  knowledge
- Open dropdown on any of the view
- Click within that view anywhere outside dropdown
- you notice that dropdown did't close

Issue:

Dropdown did't close when clicking within the embedded view.

Technical Reason:

There is a function called onWindowClick in the dropdown.js file, inside which
we get different values in 'this.ui.activeElement' and 'this.myActiveEl', so the
dropdown doesn't close.

Before this commit:

When we insert an embedded view, and the user opens any dropdown and after that,
if the user clicks within the embedded view that dropdown is not closing.

After this commit:

Removing the condition related to 'this.ui.activeElement' and 'this.myActiveEl'
means that now, In the embedded view, when opening the dropdown menu, clicking
within that view close the dropdown.

task-3717014

closes odoo/enterprise#57471

X-original-commit: 0c3ab033071f14f31850e86293d2463922389024
Related: odoo/odoo#155326
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>