[REF] spreadsheet_edition: adapt pivot commands to new format
Commands to build pivots have been reworked in order to support more than
one pivot type (for now odoo, and later on, spreadsheet). To do so, the
specific properties of the pivot command have been moved to the `payload`
property. This commit adapts the usages of INSERT_PIVOT.

Task: 3748392
Part-of: odoo/enterprise#55964