[FIX] l10n_mx_edi: allow addenda access to users with permission
Fix https://github.com/odoo/enterprise/commit/b9120ba7292a2d9cdb636bf302b22901bf0ca08f prevents an error poping when a user with no read rights would access the addenda field.
This was prevented by checking if the user is admin, which is by default the only group that can read addendas (ir.ui.views). But when the website module is installed, some new permissions rules are added for the ir.ui.view model. This means that on db's with website installed, a non admin user could access the addenda field without errors, while now they cannot anymore if they are not admins.
This fix check if the user has or has not right to ir.ui.views instead of checking if he is admin.

closes odoo/enterprise#57706

X-original-commit: 06a3c585aa54190d5637575cffa0334748eb669e
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>