[FIX] sign: fix multi-line text rendering
Before this commit when inserting long lines in multi-line text field and generating the final pdf, the multi-line did not respect its boundaries at all.
This happened because font size wasn't applied properly on the frontend, so there was a difference between fontsize inside the textarea and the fontsize when writing lines on the document which caused the lines to overflow horizontally.

This commit aims to fix that issue by removing the fixed textarea font-size css property and allow it to be set dynamically based on the zoom factor and page height.

Task: 3696347
X-original-commit: b6ae13ceee0e201379b6ac1aff187c883f68c190
Part-of: odoo/enterprise#58652