[FIX] web_studio: edit_view with a clean context
Before this commit, and because of [1], the full action context was used to create or write any
oject that edit_view would use.

This caused two types of issues:
- Checking the view integrity (_check_xml) sometimes crashed because the context keys (for modifiers)
were present and return something else than a Boolean
- creating objects with "default_" keys in the context ultimately failed, at best with a crash, at worst it
ended up creating object with the wrong data.

After this commit, we still retrieve stuff (the view, in particular) with the full context to have symmetry
with the original get_view of Studio, but we write and create things with a cleaned context.

opw-3718769
opw-3719815
opw-3703559
opw-3702620
opw-3702360
opw-3698111
opw-3698108
opw-3702084

[1]: https://github.com/odoo/enterprise/pull/54552/

closes odoo/enterprise#56556

X-original-commit: 4cee29118aa89ecaef65c5c847e77d486d7dbbe6
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>