[FIX] account_online_sync: fix stuck in fetching transaction display
In some cases, when the cron to fetch transaction is running, we first set the status of the
connection to "fetching transactions" which is displayed on the dashboard. Once the process is over,
status is changed to "done". If an error happen when contacting the provider, status is also reset
in order to not display the "fetching status" message forever.
However it can happen that an error occured when creating transaction in database (wrong
configuration inside the journal, etc) and those errors were not resetting the fetching_status
flag and were not displaying any error. Hence leaving the customer with the impression that
it was still loading transactions.
This commit fixes the problem by resetting the flag and setting the connection in error.
So if such an error happen during the cron, it won't fetch transactions and display will be clean.
Error will still be raised to the customer if he performs a manual action.

closes odoo/enterprise#58013

X-original-commit: 5a52dc0ff11cd7cb55f0c21b09509ca32ed77578
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Cedric Snauwaert <csn@odoo.com>