[FIX] room: update week when changing booking to edit
Purpose:
--------
When clicking on a booking in the sidebar while already in the frontend
form view, the week shown does not change and the booking to edit is
therefore not always visible after the click.

This commit ensures than when one clicks on a booking in the sidebar, the
current week shown is the week during which the booking takes place.

Task-3609006

Part-of: odoo/enterprise#51351