[FIX] knowledge: keep imported embed default filters
The default filter on a view was not enabled when inserting said view in
Knowledge.

How to reproduce:
- Open a view with a default filter (i.e. `CRM > My Pipeline`)
- Insert it in Knowledge

Current Behavior:
- The default filter is not enabled in the embedded view in Knowledge.

Expected Behavior:
- Every active filter should be active in the embedded view after being
  inserted, no matter if it was a default filter or not.

Technical explanation:
`_activateDefaultSearchItems` is actually a toggler, and in the specific case
where a view does not have a default favorite in knowledge, and was inserted
with a default filter enabled, the method would toggle it and disable it.

task-3742241

closes odoo/enterprise#57349

X-original-commit: 962b2f55e79ac004805b0cd3a86105b367a39cf9
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>