[FIX] web_studio: correctly export monetary and its currency_field
Create a monetary on a model. In studio, remove the currency_field for that field.
Note that the currency field would have been create automatically since commit [1]

Try to export the studio customization module.

Before this commit, there was crash, because the code that retrieves the currency_field was somewhat erroneous
and a recordset of ir.model.fields was joined to a recordset of ir.model.

After this commit, there is no crash and we export the value for the currency_field of the monetary

opw-3677016

[1]: 246477105bffd4216509b8d448b5f0b24d19d368

closes odoo/enterprise#56408

Signed-off-by: Raphael Collet <rco@odoo.com>