[FIX] Knowledge: no item overlap when no title
Issue: When adding a new Calendar Item without a
title, the item does not allocate the required
space, causing it to appear slightly under the
calendar item. Additionally, if any item is moved
to the top of the calendar, it overlaps the title
placeholder, preventing title modification.

Steps to Reproduce:

1. Install Knowledge.
2. Create a new article and select
'Build an Item Calendar'.
3. Proceed without entering a title.
4. Type anything below the created Calendar and
move it to the top of it.

Solution: Modify the 'onBuildItemCalendarBtnClick'
function to utilize a "computed" title rather than
directly using the name. This ensures that a
placeholder title is always present, allowing for
later deletion or modification.

opw-3725805

closes odoo/enterprise#57622

X-original-commit: 899804795fbb9b8794b4bda5d5ac676a864dec96
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>