[FIX] l10n_pe_edi: Add domain 01 for export invoices
In commit 355250d, the domain was adjusted to automatically select the
invoice document type for customers with a RUC and limit others to the
"boleta" document type. This generally holds true, except for export
invoices requiring the 01 document type. This commit addresses the
current restriction, allowing the selection of document type 01 other
than "boleta" or "debit note boleta."

To fix, was inherited the compute method, and extend the depends to add
the operation type and allow generate invoice for exportation cases.

closes odoo/enterprise#58464

X-original-commit: 1599e5b634f45677aa2ab452d201ce3c5c85ae32
Related: odoo/odoo#157317
Signed-off-by: Laurent Smet (las) <las@odoo.com>