[IMP] *_sale_*_renting,*_sale_subscription: use Owl product configurator in frontend
This change adapts the backend product configurator (which has been recently migrated to Owl [1])
to make it usable in the frontend as well, and uses it in the frontend.

This change also adds support for rentals and subscriptions in the product configurator, both in
the backend and the frontend (e.g. it displays the rental duration/billing period and ensures the
main and optional products use the same rental duration/billing period).

[1] https://github.com/odoo/odoo/pull/106511

task-3572125

closes odoo/enterprise#56469

Related: odoo/odoo#153779
Related: odoo/upgrade#5896
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>