[FIX] worksheet, industry_fsm_report: fix closing studio when design a
template

description :
When a user is trying to deisgn a new template, upon quitting studio
mode, an error is trigger, or in some case the studio page is not even
opened and crashes.

steps to reproduce:
on any saas-17.1 -master version
- open quality app
- select configuration then quality worksheet template
- select Design template on any worksheet
- click on the 'close' button to leave studio
an error message is displayed.

expected behavior: studio is closed correctly and the changes are saved

source of the issue:
The worksheet_id was added to the context in generic use case to handle
a fsm specific one. This was done in order to let the opportunity to
later on use the same logic on other module if needed. This id being
present in the context is triggering the fsm specific steps when it is
unneeded.

Solution:
Remove the id from the generic method 'get_x_model_from_action' and add
it in the overwrite inside the fsm module.

note:
No tests were added because the only way to test this use case is to add
a js tour, which seemed a bit overkill for such a small fix.

affected version: saas-17.1 - master

closes odoo/enterprise#56567

X-original-commit: db1ac848fb024b67fabb13fa674d9298686558e9
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>