[REF] account_reports, l10n_fr_reports: generic _vat_closing_entry_results_rounding
The tax closing in France did an override of `_postprocess_vat_closing_entry_results`
to be able to compute and compensate (by creating a move) difference from rounding
taxes.

As this functionality starts to be needed in other countries, we extract that method
to `account_generic_tax_report` and made it generic.

task-3691312

closes odoo/enterprise#59173

X-original-commit: 1d23866a8caa11a8dce46a1f93861460724e9077
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>