[FIX] room: use correct luxon locale right from loading
Purpose:
--------
Since the redesign of the frontend view (commit [1]), the locale used
when loading the frontend view is not the same as the one used once
the view is loaded.
This can be seen in the current datetime shown in the top right corner
(its format changes when it is updated the first time), or when a new
booking is added in the sidebar (it won't have the same time format as
the existing ones).

This commit ensures that the locale is correctly set when loading the
frontend view.

[1]: 4e0cde8

Task-3609006

X-original-commit: afbeeb263bd6cdd2c8d544c3a1f32243cd7c4ac4
Part-of: odoo/enterprise#56505