[FIX] account_intrastat: Fix related fields using one2many
The aim of this commit was originally adding a test to verify
that we could unfold the report. After some investigations,
we discovered that we have several fields in account_intrastat
that are using one2many fields in related fields (causing an
issue in our test). Thus, this commit is changing related fields
to computed one. We decided arbitrary that we'll take the first
value of the one2many field for each computed ones.

no task id

closes odoo/enterprise#56869

X-original-commit: 30a33f0942079f758f5859f25adc1645a1e7cd72
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>