[FIX] timehseet_grid_holidays: ensure test is not dependant of demo data
Before this commit, when the demo data adds a public leave for the current day,
the test `test_adjust_grid_holidays` could fail because we could fetch the
timesheet generating the public leave instead of the one created.

This commit adds a freeze_time on the test to be sure the current date is not
the current one but `2018-06-02`.

runbot-57194

closes odoo/enterprise#56986

X-original-commit: 0fa50fac91dbfbeb8f67ed921e47c0ee4e833567
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>