[FIX] room: prevent remaining time flickers
Purpose:
--------
Currently when a booking ends, the remaining time shown could be minus 1
second for a split second.

This commit ensures that when a booking ends, the remaining time shown is
updated at the same time as the sidebar, and does not show a negative
remaining time.

Task-3609006

X-original-commit: 2e693bbd11303db3e48120cdcc35517074a3b83d
Part-of: odoo/enterprise#56505