[FIX] room: do not show "stop" on midnight when selected
Purpose:
--------
Since the redesign of the frontend view (commit [1]), "start" and "stop"
hints are not shown in the start and end slots of the selected bookings
in the frontend form view anymore. It was still shown on the midnight
slot.
This commit removes it for the midnight slot too for consistency.

[1]: 4e0cde8

Task-3609006

X-original-commit: 56ea6412eb3f1fe34f414b57ad5f8e5e990101c2
Part-of: odoo/enterprise#56505