[FIX] room: leave form view if editing booking is deleted
Purpose:
--------
If a user is editing a booking in the frontend view and another user
deletes this booking, the user can continue to edit this booking but
an error will be raised when confirming the modification.

This commit ensures that the form view will close when the booking being
edited is deleted by someone else.

Task-3609006

X-original-commit: 5cfffb8ec789015d9fd08c33d327740b8f537102
Part-of: odoo/enterprise#56505