[FIX] web_gantt: display start and end time when pill is displayed in one day
Before this commit, when the shift has the following planned dates:
- start date: 2024-02-02 16:00:00
- end date: 2024-03-02 01:00:00
The shift is displayed in one day in the gantt view, the reason is because the
gantt view truncates the date if between the date of the end date and the end
date there is less than 3 hours than the pill will not be displayed in that day
(same logic for the date of the start date and the start date) and will only be
displayed in the other days, in that example, the shift will only be displayed
in `2024-02-02`. The problem is the custom code in planning considers the shift
in 2 days and so the start and end time are not displayed in the pill (=shift in
that case) which is a bit annoying.

This commit manages that special behavior to ensure the start and end time are
displayed in that case since the shift is only displayed in 1 day.

opw-3588166

closes odoo/enterprise#57273

X-original-commit: a0543e4ad6ca7b2a495e3534e2f0dc6ca06a57d5
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>