[FIX] documents_spreadsheet: pagination with search term
Before this fix, entering a search term in the bar caused pagination to
malfunction. The issue stemmed from the offset not resetting to 0 upon
entering a search term, leading to incorrect behavior. Additionally,
the total count failed to update with search inputs.

This commit resolves the issue by resetting the offset to 0 upon search
term entry and updating the total count using the domain passed to
the _fetchPagerTotal method.

Task ID: 3631981

closes odoo/enterprise#56789

X-original-commit: c54df170b596703078f7196dd3e5527da526d264
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>