[FIX] documents_spreadsheet: pagination with search term
Before this fix, entering a search term in the bar caused pagination to
malfunction. The issue stemmed from the offset not resetting to 0 upon
entering a search term, leading to incorrect behavior. Additionally,
the total count failed to update with search inputs.

This commit resolves the issue by resetting the offset to 0 upon search
term entry and updating the total count using the domain passed to
the _fetchPagerTotal method.

Task ID: 3631981

closes odoo/enterprise#56669

X-original-commit: 2c4c6bb12b99bb8e9f92fa56f4ef559095021262
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>