[IMP] account_online_synchronization: Send totpenabled
One of our future bank synchronization provider, bLink,
is asking us to make sure that customers can only connect
to their banks if they have a 2FA to login on their database.
As we don't want to enforce this behavior for other providers,
we decide to send the value when we open the bank selection view.
If a customer tries to connect with a bLink institution, an error
will be sent by Odoo Fin proxy telling that connection with bLink
need a 2FA enabled.
NB: This commit is only for sending the info when opening the iframe,
all the logic is handled by Odoo Fin.

closes odoo/enterprise#59119

Task-id: 3637581
X-original-commit: 181cf963b53f0859b60fcbbbbaadcde5a65c4085
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>