[FIX] appointment: show selected resource for resource_time
For resource appointments with Resource Then Time
setting, when clicking on a day to select a slot,
for instance on update of the slots when changing
users, or when simply when selecting a day, the
resource_selected_id is not added properly to the
url as it should, hence not showing the selection
in the details column in the /info screen.

This is only due to not using the value of the
select for the resource. Fix the issue by using
the value of the resource dropdown if any.

Task-3603354

closes odoo/enterprise#56585

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>