[FIX] appointment: allow resource selection for time_resource
When coming back to the time selection screen, from an
error or through chevrons, the previously chosen resource
(resource_selected_id in the url) was selected for 'Time
then resource' in the resource dropdown, and at loading,
also when computing available slots.

In order to let the user choose their resource in that case,
do not use the selected resource for time_resource types, nor
for computing availabilities at first loading, nor to restrict
the resource dropdown.

Task-3603354

closes odoo/enterprise#58240

X-original-commit: 0f625ce789817cd6fb9594e2f636c298440fae48
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Noé Antoine (nan) <nan@odoo.com>