[FIX] web_studio: forbid properties field in related selection.
On a view, drag and drop a new related field. Go into one m2o field.

Before this commit, it was possible to select a field of type "properties".
This is not supported by python at the time.
To do so, one would have to reflect the property "definition" of the python field.
Also, some code would have to be written to create that other field, related as well
that'd contain the properties' definition.

After this commit, to make things simpler, we just blacklist property fields in the related
field selector.

opw-3730900

closes odoo/enterprise#56790

X-original-commit: b3e6d0d63caf4b06f1c2847db37a04f822f37f86
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>