[FIX] appointment: fix default resource calendar
Add and use a default resource calendar encompassing every hours of every days.
This allows to take into account leaves for resource easily as we need
a resource calendar for the computation. Indeed when the resource calendar
was removed, the one set on the company was used in place which led to
weird result for the end user as he didn't expect a working schedule after
removing the one set on the resource.

X-original-commit: 3c0b560995859ebb0fe11f8889a5c7fc0d6d609a

task-3562191

X-original-commit: 6cb5c6ee0f0c66e87e0cc89746dbc90817bc384a
Part-of: odoo/enterprise#57438