[FIX] l10n_ec_edi: missing edi information fields
**Current behavior:**
Creating an EDI document from an invoice does not have fields
for currency (moneda) in some places where it is expected.

**Expected behavior:**
Creating an EDI document from an invoice will include these
fields in all their expected places.

**Steps to reproduce:**
0. Install the l10n_ec_edi accounting module; switch current
     company to 'EC Company'

1. Duplicate an existing invoice and validate it

2. Process the invoice to receive an XML edi document

3. This form is purportedly missing expected fields

**Cause of the issue:**
The currency information is not set in a manner to be
accessed in all the places it is needed in edi_document.xml.

**Fix:**
Currency information is passed to the requisite scope via
methods in the l10n_ec_edi account.move class.

opw-3669363

closes odoo/enterprise#57912

X-original-commit: 79551fbcd548297400af81dc2d6516e9462101e2
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Vincent Ethan <etvi@odoo.com>