[FIX] l10n_lt_reports: fix hierarchy levels
In the Balance Sheet:
The headers "TOTAL ASSETS" and "TOTAL EQUITY AND LIABILITIES" should
be on the same level.

In the Profit & Loss:
The headers "Net turnover" and "Cost of sales" should
be on the same level as the rest of the headers.

This is corrected in this commit.

The sepcs can be found in task-3701151.

closes odoo/enterprise#57433

X-original-commit: 9109ed1cc28ef035a67d266dfd8c963a286eb2fc
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>