[FIX] l10n_lt_reports: fix hierarchy levels
In the Balance Sheet:
The headers "TOTAL ASSETS" and "TOTAL EQUITY AND LIABILITIES" should
be on the same level.

In the Profit & Loss:
The headers "Net turnover" and "Cost of sales" should
be on the same level as the rest of the headers.

This is corrected in this commit.

The sepcs can be found in task-3701151.

closes odoo/enterprise#57481

X-original-commit: e8c11fde8876d90563db7f01940cbcd5bed949a9
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>