[IMP] timesheet_grid: keep timer between views
Remove unnecessary code from `timesheet_timer_hooks.js`.
Add a hook `onRecordChanged` to the timesheet controllers to manage the
usecase where we start a timer, but no "default" project is found.

In that case, we create a "fake" record in the view (we don't save it
because without project, it is not qualified as a timesheet but as an
analytic line). As soon as a project is set on it however, we want to
save it, and that is what the hook does.

related: https://github.com/odoo/odoo/pull/154134
task-3594778

Part-of: odoo/enterprise#56679
Related: odoo/odoo#154134
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>