[IMP] planning: add an occupancy percentage over time on gantt view
In the gantt view, when grouping by resource > something else, this commit
introduces an occupancy percentage that is displayed next to the aggregated
allocated hours of each pill. This percentage actually represents the same
information as the colors added in https://github.com/odoo/enterprise/pull/38343
but in a textual way.
Note that the occupancy percentage is only displayed for the week scale.

occupancy rate = total number of allocated hours for a given period of time /
total number of hours to work for that period of time (according to the resource
working calendar or the company calendar in case of flexible hours)

task-3651242

closes odoo/enterprise#56678

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>